+++ frontend/dev-dist/sw.js
@@ -81,7 +81,7 @@ define(['./workbox-d3a36d94'], (function (workbox) { 'use strict';
     "revision": "3ca0b8505b4bec776b69afdba2768812"
   }, {
     "url": "/index.html",
-    "revision": "0.1b560k9t7qg"
+    "revision": "0.jqv9brlsen"
   }], {});
   workbox.cleanupOutdatedCaches();
   workbox.registerRoute(new workbox.NavigationRoute(workbox.createHandlerBoundToURL("/index.html"), {